modify
class MemosData　作成 h をerb側に入れる
