+++ views/edit.erb
@@ -1,11 +1,11 @@
-<form method="post" action='/memos/<%= @memo["id"] %>' >
+<form method="post" action='/memos/<%=h @memo["id"] %>' >
  <input id="hidden" type="hidden" name="_method" value="patch">
 <div>
  <label for="title">題名:</label>
- <input type="text" name="title"  value=<%= @memo["title"] %>>
+ <input type="text" name="title"  value=<%=h @memo["title"] %>>
  <br>
  <label for="content">内容:</label>
- <textarea name="content" rows="10" cols="80" ><%= @memo["content"] %></textarea>
+ <textarea name="content" rows="10" cols="80" ><%=h @memo["content"] %></textarea>
  <br>
 <button type="submit" > 変更</button>
 </div>
